Solve sigmoid-numpy

+++ sigmoid-numpy/sigmoid-numpy.py
@@ -4,6 +4,6 @@ def sigmoid(x):
     """
     Vectorized sigmoid function.
     """
-    x = np.asarray(x , dtype = float)
+    x = np.asarray(x )
     prob  = 1/ ( 1 + np.exp(-x))
     return prob
